Version 0.8 beta, used on kasteria.com

+++ otclient/src/framework/graphics/atlas.h
@@ -30,7 +30,7 @@ private:
 
     FrameBufferPtr m_atlas[2];
     std::map<TexturePtr, CachedAtlasObject> m_textures[2];
-    std::map<size_t, CachedAtlasObject> m_outfits[2];
+    std::map<uint64_t, CachedAtlasObject> m_outfits[2];
     std::list<Point> m_locations[2][5];
     size_t m_size;
 };
